move `resolve_snref()` from `utils.py` to `odxlink.py`
As usual, this was suggested by [at]kayoub5. Thanks!
